Split up forms routes test file
