+++ src/routes/forms.test.js
@@ -1,37 +1,12 @@
-import {
-  ConditionType,
-  Coordinator,
-  FormStatus,
-  OperatorName,
-  organisations
-} from '@defra/forms-model'
+import { FormStatus, organisations } from '@defra/forms-model'
 import Boom from '@hapi/boom'
 
-import {
-  buildCondition,
-  buildDefinition,
-  buildList,
-  buildQuestionPage,
-  buildSummaryPage,
-  buildTextFieldComponent
-} from '~/src/api/forms/__stubs__/definition.js'
 import { FormAlreadyExistsError } from '~/src/api/forms/errors.js'
-import {
-  createComponentOnDraftDefinition,
-  deleteComponentOnDraftDefinition,
-  updateComponentOnDraftDefinition
-} from '~/src/api/forms/service/component.js'
-import {
-  addConditionToDraftFormDefinition,
-  removeConditionOnDraftFormDefinition,
-  updateConditionOnDraftFormDefinition
-} from '~/src/api/forms/service/conditions.js'
 import {
   createDraftFromLive,
   createLiveFromDraft,
   getFormDefinition,
-  listForms,
-  reorderDraftFormDefinitionPages
+  listForms
 } from '~/src/api/forms/service/definition.js'
 import {
   createForm,
@@ -40,17 +15,7 @@ import {
   removeForm,
   updateFormMetadata
 } from '~/src/api/forms/service/index.js'
-import {
-  addListToDraftFormDefinition,
-  removeListOnDraftFormDefinition,
-  updateListOnDraftFormDefinition
-} from '~/src/api/forms/service/lists.js'
 import { migrateDefinitionToV2 } from '~/src/api/forms/service/migration.js'
-import {
-  createPageOnDraftDefinition,
-  deletePageOnDraftDefinition,
-  patchFieldsOnDraftDefinitionPage
-} from '~/src/api/forms/service/page.js'
 import { createServer } from '~/src/api/server.js'
 import { auth } from '~/test/fixtures/auth.js'
 
@@ -60,7 +25,6 @@ jest.mock('~/src/api/forms/service/definition.js')
 jest.mock('~/src/api/forms/service/page.js')
 jest.mock('~/src/api/forms/service/component.js')
 jest.mock('~/src/api/forms/service/migration.js')
-jest.mock('~/src/api/forms/service/lists.js')
 jest.mock('~/src/api/forms/service/conditions.js')
 
 describe('Forms route', () => {
@@ -82,8 +46,6 @@ describe('Forms route', () => {
   const internalErrorStatusCode = 500
   const jsonContentType = 'application/json'
   const id = '661e4ca5039739ef2902b214'
-  const pageId = 'c7b9f0fa-3223-46b8-b7d3-b2bf00f37155'
-  const componentId = '64d9c012-2238-4ab2-ab11-6290d3c0cf15'
   const now = new Date()
   const authorId = 'f50ceeed-b7a4-47cf-a498-094efc99f8bc'
   const authorDisplayName = 'Enrique Chase'
@@ -101,17 +63,6 @@ describe('Forms route', () => {
     organisation: 'Defra',
     teamName: 'Defra Forms',
     teamEmail: 'defraforms@defra.gov.uk'
-  }
-
-  const stubTextFieldComponent = buildTextFieldComponent({
-    id: undefined,
-    title: 'What is your name?',
-    name: 'Ghcbmw'
-  })
-
-  /** @satisfies {PatchPageFields} */
-  const stubPatchPageFields = {
-    title: 'Updated title for page'
   }
 
   /**
@@ -606,404 +557,6 @@ describe('Forms route', () => {
         status: 'created-draft'
       })
     })
-
-    test('Testing POST /forms/{id}/definition/draft/pages adds a new page to the db and populates id', async () => {
-      const pagePayload = buildQuestionPage({
-        id: undefined
-      })
-      const createPageMock = jest.mocked(createPageOnDraftDefinition)
-      createPageMock.mockResolvedValue(pagePayload)
-
-      const response = await server.inject({
-        method: 'POST',
-        url: `/forms/${id}/definition/draft/pages`,
-        payload: pagePayload,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual(pagePayload)
-      expect(createPageMock).toHaveBeenCalledWith(
-        id,
-        { ...pagePayload, id: expect.any(String) },
-        { ...author, id: expect.any(String) }
-      )
-    })
-
-    test('Testing PATCH /forms/{id}/definition/draft/pages/{pageId} updates fields on a page', async () => {
-      const questionPage = buildQuestionPage({
-        title: 'Updated title for page'
-      })
-      const patchFieldsOnDraftPageMock = jest
-        .mocked(patchFieldsOnDraftDefinitionPage)
-        .mockResolvedValue(questionPage)
-
-      const response = await server.inject({
-        method: 'PATCH',
-        url: `/forms/${id}/definition/draft/pages/${pageId}`,
-        payload: stubPatchPageFields,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual(questionPage)
-      const [calledFormId, calledPageId, patchedFields] =
-        patchFieldsOnDraftPageMock.mock.calls[0]
-      expect([calledFormId, calledPageId, patchedFields]).toEqual([
-        id,
-        pageId,
-        stubPatchPageFields
-      ])
-    })
-
-    test('Testing POST /forms/{id}/definition/draft/pages/order reorders the pages in the db', async () => {
-      const pageOneId = '5113a8ab-b297-46b5-b732-7fe42660c4db'
-      const pageTwoId = 'd3dc6af2-3235-4455-80f7-941f0bf69c4f'
-
-      const expectedDefinition = buildDefinition({
-        pages: [
-          buildQuestionPage({
-            id: pageOneId,
-            title: 'Page one'
-          }),
-          buildQuestionPage({
-            id: pageOneId,
-            title: 'Page two'
-          }),
-          buildSummaryPage()
-        ]
-      })
-      jest
-        .mocked(reorderDraftFormDefinitionPages)
-        .mockResolvedValue(expectedDefinition)
-
-      const response = await server.inject({
-        method: 'POST',
-        url: `/forms/${id}/definition/draft/pages/order`,
-        payload: [pageOneId, pageTwoId],
-        auth
-      })
-
-      expect(response.result).toEqual(expectedDefinition)
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-    })
-
-    test('Testing POST /forms/{id}/definition/draft/pages/{pageId}/components adds a new component to a page', async () => {
-      const expectedComponent = buildTextFieldComponent({
-        ...stubTextFieldComponent,
-        id: '3813a55d-0958-47f9-8522-94b3fc3818d7'
-      })
-      const createComponentOnDraftDefinitionMock = jest
-        .mocked(createComponentOnDraftDefinition)
-        .mockResolvedValue(expectedComponent)
-
-      const response = await server.inject({
-        method: 'POST',
-        url: `/forms/${id}/definition/draft/pages/${pageId}/components`,
-        payload: stubTextFieldComponent,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual(expectedComponent)
-      const [calledFormId, calledPageId, component, , prepend] =
-        createComponentOnDraftDefinitionMock.mock.calls[0]
-      expect([calledFormId, calledPageId, component, prepend]).toEqual([
-        id,
-        pageId,
-        {
-          ...stubTextFieldComponent,
-          id: expect.any(String)
-        },
-        false
-      ])
-    })
-
-    test('Testing PUT /forms/{id}/definition/draft/pages/{pageId}/components/{componentId} updates a component on a page', async () => {
-      const updatedComponent = buildTextFieldComponent({
-        id: componentId,
-        title: 'New component title'
-      })
-      const updateComponentOnDraftPageMock = jest
-        .mocked(updateComponentOnDraftDefinition)
-        .mockResolvedValue(updatedComponent)
-
-      const response = await server.inject({
-        method: 'PUT',
-        url: `/forms/${id}/definition/draft/pages/${pageId}/components/${componentId}`,
-        payload: updatedComponent,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual(updatedComponent)
-      expect(updateComponentOnDraftPageMock).toHaveBeenCalled()
-      const [calledFormId, calledPageId, calledComponentId, component] =
-        updateComponentOnDraftPageMock.mock.calls[0]
-      expect([
-        calledFormId,
-        calledPageId,
-        calledComponentId,
-        component
-      ]).toEqual([id, pageId, componentId, updatedComponent])
-    })
-
-    test('Testing POST /forms/{id}/definition/draft/pages/{pageId}/components?prepend=true adds a new component to the start of a page', async () => {
-      const expectedComponent = buildTextFieldComponent({
-        ...stubTextFieldComponent,
-        id: '3813a55d-0958-47f9-8522-94b3fc3818d7'
-      })
-      const createComponentOnDraftDefinitionMock = jest
-        .mocked(createComponentOnDraftDefinition)
-        .mockResolvedValue(expectedComponent)
-
-      const response = await server.inject({
-        method: 'POST',
-        url: `/forms/${id}/definition/draft/pages/${pageId}/components?prepend=true`,
-        payload: stubTextFieldComponent,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      const [, , , , prepend] =
-        createComponentOnDraftDefinitionMock.mock.calls[0]
-      expect(prepend).toBe(true)
-    })
-
-    test('Testing POST /forms/{id}/definition/draft/lists', async () => {
-      const payload = buildList({
-        id: undefined
-      })
-
-      const expectedList = {
-        ...buildList({
-          ...payload
-        }),
-        id: '9719c91f-4341-4dc8-91a5-cab7bbdddb83'
-      }
-
-      const createComponentOnDraftDefinitionMock = jest
-        .mocked(addListToDraftFormDefinition)
-        .mockResolvedValue(expectedList)
-
-      const response = await server.inject({
-        method: 'POST',
-        url: `/forms/${id}/definition/draft/lists`,
-        payload,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual({
-        id: '9719c91f-4341-4dc8-91a5-cab7bbdddb83',
-        list: expectedList,
-        status: 'created'
-      })
-      const [, list] = createComponentOnDraftDefinitionMock.mock.calls[0]
-      expect(list).toEqual({
-        ...expectedList,
-        id: expect.any(String)
-      })
-    })
-
-    test('Testing PUT /forms/{id}/definition/draft/lists/{listId}', async () => {
-      const listId = '9719c91f-4341-4dc8-91a5-cab7bbdddb83'
-      const list = buildList({
-        id: '9719c91f-4341-4dc8-91a5-cab7bbdddb83'
-      })
-
-      const updateList = jest
-        .mocked(updateListOnDraftFormDefinition)
-        .mockResolvedValue(list)
-
-      const response = await server.inject({
-        method: 'PUT',
-        url: `/forms/${id}/definition/draft/lists/${listId}`,
-        payload: list,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual({
-        id: '9719c91f-4341-4dc8-91a5-cab7bbdddb83',
-        list,
-        status: 'updated'
-      })
-      const [, calledId, calledList] = updateList.mock.calls[0]
-      expect(calledId).toEqual(listId)
-      expect(calledList).toEqual(list)
-    })
-
-    test('Testing DELETE /forms/{id}/definition/draft/lists/{listId}', async () => {
-      const listId = '9719c91f-4341-4dc8-91a5-cab7bbdddb83'
-
-      const response = await server.inject({
-        method: 'DELETE',
-        url: `/forms/${id}/definition/draft/lists/${listId}`,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual({
-        id: listId,
-        status: 'deleted'
-      })
-      const [calledFormId, calledId] = jest.mocked(
-        removeListOnDraftFormDefinition
-      ).mock.calls[0]
-      expect(calledFormId).toEqual(id)
-      expect(calledId).toEqual(listId)
-    })
-
-    test('Testing DELETE /forms/{id}/definition/draft/pages/{pageId}', async () => {
-      const pageId = '9719c91f-4341-4dc8-91a5-cab7bbdddb83'
-
-      const response = await server.inject({
-        method: 'DELETE',
-        url: `/forms/${id}/definition/draft/pages/${pageId}`,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual({
-        id: pageId,
-        status: 'deleted'
-      })
-      const [calledFormId, calledId] = jest.mocked(deletePageOnDraftDefinition)
-        .mock.calls[0]
-      expect(calledFormId).toEqual(id)
-      expect(calledId).toEqual(pageId)
-    })
-
-    test('Testing POST /forms/{id}/definition/draft/conditions', async () => {
-      const payload = buildCondition({
-        id: undefined,
-        coordinator: Coordinator.AND,
-        items: [
-          {
-            id: '6746b15f-69f9-454c-a324-c62420069618',
-            componentId: '99cfccef-6292-4e7f-895f-4404751beb3c',
-            operator: OperatorName.Is,
-            value: {
-              type: ConditionType.StringValue,
-              value: 'Enrique Chase'
-            }
-          },
-          {
-            id: '018a7d26-d1f1-4bc6-bf48-b882b69bfbaa',
-            conditionId: '70de3a17-d29c-45e4-a00b-3ce0d05a8ed2'
-          }
-        ]
-      })
-
-      const expectedCondition = {
-        ...buildCondition({
-          ...payload
-        }),
-        id: '295e6e7a-5388-4f3f-9d99-1a1e749b6b0c'
-      }
-
-      const addConditionToDraftFormDefinitionMock = jest
-        .mocked(addConditionToDraftFormDefinition)
-        .mockResolvedValue(expectedCondition)
-
-      const response = await server.inject({
-        method: 'POST',
-        url: `/forms/${id}/definition/draft/conditions`,
-        payload,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual({
-        id: '295e6e7a-5388-4f3f-9d99-1a1e749b6b0c',
-        condition: expectedCondition,
-        status: 'created'
-      })
-      const [, condition] = addConditionToDraftFormDefinitionMock.mock.calls[0]
-      expect(condition).toEqual({
-        ...expectedCondition,
-        id: expect.any(String)
-      })
-    })
-
-    test('Testing PUT /forms/{id}/definition/draft/conditions/{conditionId}', async () => {
-      const conditionId = '6e4c2f74-5bd9-48b4-b991-f2a021dcde59'
-      const condition = buildCondition({
-        id: conditionId,
-        coordinator: Coordinator.AND,
-        items: [
-          {
-            id: '6746b15f-69f9-454c-a324-c62420069618',
-            componentId: '99cfccef-6292-4e7f-895f-4404751beb3c',
-            operator: OperatorName.Is,
-            value: {
-              type: ConditionType.StringValue,
-              value: 'Enrique Chase'
-            }
-          },
-          {
-            id: '018a7d26-d1f1-4bc6-bf48-b882b69bfbaa',
-            conditionId: '70de3a17-d29c-45e4-a00b-3ce0d05a8ed2'
-          }
-        ]
-      })
-
-      const updateCondition = jest
-        .mocked(updateConditionOnDraftFormDefinition)
-        .mockResolvedValue(condition)
-
-      const response = await server.inject({
-        method: 'PUT',
-        url: `/forms/${id}/definition/draft/conditions/${conditionId}`,
-        payload: condition,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual({
-        id: '6e4c2f74-5bd9-48b4-b991-f2a021dcde59',
-        condition,
-        status: 'updated'
-      })
-      const [, calledId, calledCondition] = updateCondition.mock.calls[0]
-      expect(calledId).toEqual(conditionId)
-      expect(calledCondition).toEqual(condition)
-    })
-
-    test('Testing DELETE /forms/{id}/definition/draft/conditions/{conditionId}', async () => {
-      const conditionId = '9719c91f-4341-4dc8-91a5-cab7bbdddb83'
-
-      const response = await server.inject({
-        method: 'DELETE',
-        url: `/forms/${id}/definition/draft/conditions/${conditionId}`,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual({
-        id: conditionId,
-        status: 'deleted'
-      })
-      const [calledFormId, calledId] = jest.mocked(
-        removeConditionOnDraftFormDefinition
-      ).mock.calls[0]
-      expect(calledFormId).toEqual(id)
-      expect(calledId).toEqual(conditionId)
-    })
   })
 
   describe('Error responses', () => {
@@ -1270,221 +823,6 @@ describe('Forms route', () => {
       expect(response.result).toMatchObject({
         error: 'FormAlreadyExistsError',
         message: 'Form with slug my-title already exists'
-      })
-    })
-
-    test('Testing POST /forms/{id}/definition/draft/pages with invalid payload returns validation errors', async () => {
-      const invalidPageObject /** @type {Page} */ = buildQuestionPage({
-        id: 'not-a-valid-id',
-        path: '/status'
-      })
-
-      const response = await server.inject({
-        method: 'POST',
-        url: `/forms/${id}/definition/draft/pages`,
-        payload: invalidPageObject,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(badRequestStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toMatchObject({
-        error: 'Bad Request',
-        message: '"path" contains an invalid value. "id" must be a valid GUID',
-        statusCode: 400,
-        validation: {
-          keys: ['path', 'id'],
-          source: 'payload'
-        }
-      })
-    })
-
-    test('Testing POST /forms/{id}/definition/draft/pages/order with invalid payload returns validation errors', async () => {
-      const response = await server.inject({
-        method: 'POST',
-        url: `/forms/${id}/definition/draft/pages/order/`,
-        payload: ['not-a-valid-uuid'],
-        auth
-      })
-
-      expect(response.statusCode).toEqual(badRequestStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toEqual({
-        error: 'Bad Request',
-        message:
-          '"[0]" must be a valid GUID. "value" does not contain 1 required value(s)',
-        statusCode: 400,
-        validation: {
-          keys: ['0', ''],
-          source: 'payload'
-        },
-        custom: {
-          defaultError: expect.anything()
-        }
-      })
-    })
-
-    test('Testing POST /forms/{id}/definition/draft/pages/{pageId} with invalid payload returns validation errors', async () => {
-      const invalidPatchPayload = {
-        id: 1234
-      }
-      const response = await server.inject({
-        method: 'PATCH',
-        url: `/forms/${id}/definition/draft/pages/${pageId}`,
-        payload: invalidPatchPayload,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(badRequestStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toMatchObject({
-        error: 'Bad Request',
-        message: '"id" is not allowed',
-        statusCode: 400,
-        validation: {
-          keys: ['id']
-        }
-      })
-    })
-
-    const invalidComponent = buildTextFieldComponent({
-      id: 'not-a-valid-id'
-    })
-
-    test.each([
-      {
-        url: `/forms/${id}/definition/draft/pages/not-a-valid-guid/components`,
-        errors: {
-          message: '"pageId" must be a valid GUID',
-          validation: {
-            keys: ['pageId'],
-            source: 'params'
-          }
-        }
-      },
-      {
-        url: `/forms/${id}/definition/draft/pages/${pageId}/components`,
-        errors: {
-          message: '"id" must be a valid GUID',
-          validation: {
-            keys: ['id'],
-            source: 'payload'
-          }
-        }
-      }
-    ])(
-      'Testing POST /forms/{id}/definition/draft/pages/{pageId}/components with invalid payload returns validation errors',
-      async ({ url, errors }) => {
-        const response = await server.inject({
-          method: 'POST',
-          url,
-          payload: invalidComponent,
-          auth
-        })
-
-        expect(response.statusCode).toEqual(badRequestStatusCode)
-        expect(response.headers['content-type']).toContain(jsonContentType)
-        expect(response.result).toMatchObject({
-          error: 'Bad Request',
-          message: errors.message,
-          statusCode: 400,
-          validation: errors.validation
-        })
-      }
-    )
-
-    test('Testing POST /forms/{id}/definition/draft/lists with invalid payload returns validation errors', async () => {
-      const invalidListPayload = buildList({
-        // @ts-expect-error invalid parameter
-        unknown: 1
-      })
-      const response = await server.inject({
-        method: 'POST',
-        url: `/forms/${id}/definition/draft/lists`,
-        payload: invalidListPayload,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(badRequestStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toMatchObject({
-        error: 'Bad Request',
-        message: '"unknown" is not allowed',
-        statusCode: 400,
-        validation: {
-          keys: ['unknown']
-        }
-      })
-    })
-
-    test('Testing PUT /forms/{id}/definition/draft/lists/{listId} with invalid payload returns validation errors', async () => {
-      const invalidListPayload = buildList({
-        id: undefined
-      })
-      const response = await server.inject({
-        method: 'PUT',
-        url: `/forms/${id}/definition/draft/lists/8d05e220-2145-40f4-9508-fe946dec6fd9`,
-        payload: invalidListPayload,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(badRequestStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toMatchObject({
-        error: 'Bad Request',
-        message: '"id" is required',
-        statusCode: 400,
-        validation: {
-          keys: ['id']
-        }
-      })
-    })
-
-    test('Testing PUT /forms/{id}/definition/draft/pages/{pageId}/component/{componentId} with invalid payload returns validation errors', async () => {
-      const invalidPatchPayload = {
-        id: 1234
-      }
-      const response = await server.inject({
-        method: 'PUT',
-        url: `/forms/${id}/definition/draft/pages/${pageId}/components/${componentId}`,
-        payload: invalidPatchPayload,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(badRequestStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toMatchObject({
-        error: 'Bad Request',
-        message: '"type" is required. "id" must be a string',
-        statusCode: 400,
-        validation: {
-          keys: ['type', 'id']
-        }
-      })
-    })
-
-    test('Testing PUT /forms/{id}/definition/draft/pages/{pageId}/component/{componentId} without an id returns validation errors', async () => {
-      const componentWithoutAnId = buildTextFieldComponent({
-        title: 'New component title'
-      })
-      delete componentWithoutAnId.id
-
-      const response = await server.inject({
-        method: 'PUT',
-        url: `/forms/${id}/definition/draft/pages/${pageId}/components/${componentId}`,
-        payload: componentWithoutAnId,
-        auth
-      })
-
-      expect(response.statusCode).toEqual(badRequestStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toMatchObject({
-        error: 'Bad Request',
-        message: '"id" is required',
-        statusCode: 400,
-        validation: {
-          keys: ['id']
-        }
       })
     })
 
@@ -1780,36 +1118,10 @@ describe('Forms route', () => {
         }
       })
     })
-
-    test('Testing DELETE /forms/{id}/pages/{pageId}/components/{componentId} route returns a "deleted" status', async () => {
-      const deleteComponentOnDraftDefinitionMock = jest
-        .mocked(deleteComponentOnDraftDefinition)
-        .mockResolvedValue()
-
-      const response = await server.inject({
-        method: 'DELETE',
-        url: `/forms/${id}/definition/draft/pages/${pageId}/components/${componentId}`,
-        auth,
-        payload: {}
-      })
-
-      expect(response.statusCode).toEqual(okStatusCode)
-      expect(response.headers['content-type']).toContain(jsonContentType)
-      expect(response.result).toMatchObject({
-        componentId,
-        status: 'deleted'
-      })
-      expect(deleteComponentOnDraftDefinitionMock).toHaveBeenCalledWith(
-        id,
-        pageId,
-        componentId,
-        expect.anything()
-      )
-    })
   })
 })
 
 /**
- * @import { FormDefinition, FormMetadata, FormMetadataAuthor, FormMetadataInput, FilterOptions, PageStart, TextFieldComponent, PatchPageFields } from '@defra/forms-model'
+ * @import { FormDefinition, FormMetadata, FormMetadataAuthor, FormMetadataInput, FilterOptions } from '@defra/forms-model'
  * @import { Server } from '@hapi/hapi'
  */
